requirements.txt
streamlit
Pillow
python-dotenv
requests
spotipy
fastapi
uvicorn[standard]
python-multipart
